[ui] Display all selected statistical units in search result table
- This PR also extract the search result table row as a separate component in search/components
